backend/requirements.txt
Django==2.2.16
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn==20.0.4
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.1
Pillow==8.3.1